pyproject.toml
[tool.poetry]
name = "Recorder"
version = "0.1.0"
description = ""
authors = ["LJ Li <liliangjya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.87.0"
uvicorn = "^0.19.0"
python-decouple = "^3.6"
beanie = "^1.15.4"
email-validator = "^1.3.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
python-multipart = "^0.0.5"


[tool.poetry.group.dev.dependencies]
requests = "^2.28.1"
httpx = "^0.23.1"
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
include = ["./src"]

[tool.pytest.ini_options]
pythonpath = [
    ".", "src"
]
[tool.bandit]
exclude_dirs = ["tests"]